+++ src/common/js/index.js
@@ -3,9 +3,8 @@
  */
 
 import './constants.js';
-import './config.js';
+import './configure.js';
 //
-import './defaults.js';
 import './functions.js';
 import './i18n.js';
 import './startup.js';